## Context

Bring the Paystack provider branch onto current Lago main before applying the remaining review fixes.

## Description

Merge upstream main and regenerate the GraphQL schemas from the merged Ruby definitions.

Schema generation is repeatable. Rails eager loading and RuboCop pass. The Paystack baseline has 110 examples with 23 failures caused by missing typed-result fields; these remain for the review-fix commit.

+++ spec/graphql/types/data_exports/credit_notes/filters_input_spec.rb
@@ -6,8 +6,8 @@ RSpec.describe Types::DataExports::CreditNotes::FiltersInput do
   subject { described_class }
 
   it do
-    expect(subject).to accept_argument(:amount_from).of_type("Int")
-    expect(subject).to accept_argument(:amount_to).of_type("Int")
+    expect(subject).to accept_argument(:amount_from).of_type("BigInt")
+    expect(subject).to accept_argument(:amount_to).of_type("BigInt")
     expect(subject).to accept_argument(:billing_entity_ids).of_type("[ID!]")
     expect(subject).to accept_argument(:credit_status).of_type("[CreditNoteCreditStatusEnum!]")
     expect(subject).to accept_argument(:currency).of_type("CurrencyEnum")
